Update some scenarios enumeration
