+++ e2e-playwright/tests/EP016.spec.ts
@@ -1,57 +1,37 @@
-import { test, expect } from "@playwright/test";
-import { takeScreenshot } from "../util/util";
+import test, { expect } from "@playwright/test";
 import { LoginPage } from "../page/LoginPage";
-import { PagesEditor } from "../page/PagesEditor";
-import { Urls, URL } from "../../shared/config";
+import { TagPage } from "../page/TagPage";
+import { takeScreenshot } from "../util/util";
 
 
-test("EP016 Create new page and edit it", async ({ page }) => {
+
+/*
+    Test Case: EP016 - Verify create new Tag
+*/
+test("EP016 - Verify create new @Tag", async ({ page }) => {
     const loginPage = new LoginPage(page);
-    const pagesEditor = new PagesEditor(page);
+    const tagPage = new TagPage(page);
 
-    // Chek if user is logged in
+
+    // Given: User is logged in
     await loginPage.open();
     await loginPage.login();
+    // And Navigate to the tag page
+    await tagPage.open();
+    // When I fill the tag name
+    await tagPage.fillTagName('Test Tag');
 
-    // Navigate to Pages Editor and Create a new Page
-    await pagesEditor.open();
-    let newPageName = "Prueba edicion contenido";
-    await pagesEditor.createTestPage(newPageName);
-
-    // Screenshot the inicial state of the new page
-    await page.goto(`${URL}/prueba-edicion-contenido`, { waitUntil: "load" });
-    await takeScreenshot(page);
-
-    // Add new section
-    await page.goto(Urls.listPage, { waitUntil: "load" });
-    await page.waitForTimeout(500);
-    await pagesEditor.editPage(newPageName);
-    let newTitle = "Este título se creó editando la página";
-    let newParagraph = "Este parrafo se creó editando la página";
-    let lastParagraph = page.locator(".kg-prose").first();
-    await lastParagraph.click();
-    await page.keyboard.press("Control+ArrowDown");
-    await page.keyboard.press("Enter");
-    await page.keyboard.type("## " + newTitle);
-    await page.keyboard.press("Enter");
-    await page.keyboard.type(newParagraph);
-    await page.click("text='Update'");
-    await page.waitForTimeout(1000)
-
-
-    // verify if new page contains the new section 
-    await page.goto(`${URL}/prueba-edicion-contenido/`, { waitUntil: "load" });
-    await takeScreenshot(page);
-    let container = page.locator('.gh-content');
-    let contenidos = await container.allInnerTexts();
-
-    let titleExists = false;
-    let paragraphExists = false;
-    for (let i = 0; i < contenidos.length; i++) {
-        if (contenidos[i] == newTitle) { titleExists = true };
-        if (contenidos[i] == newParagraph) { paragraphExists = true };
+    async function getSaveTagResponse() {
+        const responsePromise = await page.waitForResponse(async (response) => {
+            if (!response.url().includes('tags')) return false;
+            return response.status() === 201 || response.status() === 200;
+        });
+        return responsePromise.json();
     }
-    expect(titleExists).toBeTruthy;
-    expect(paragraphExists).toBeTruthy;
-
+    // And I save the tag
+    await tagPage.saveTag();
+    const response = await getSaveTagResponse();
+    await takeScreenshot(page);
+    // Then It should create the tag and return the name
+    expect(response.tags[0].name).toBe('Test Tag');
 });
