+++ e2e-playwright/tests/EP018.spec.ts
@@ -4,34 +4,26 @@ import { TagPage } from "../page/TagPage";
 import { takeScreenshot } from "../util/util";
 
 
-
 /*
-    Test Case: EP018 - Verify create new Tag
+    Test Case: EP018 - Tag description should be less than 500 characters
 */
-test("EP018 - Verify create new @Tag", async ({ page }) => {
+test("EP018 - Verify tag description limit", async ({ page }) => {
     const loginPage = new LoginPage(page);
     const tagPage = new TagPage(page);
-
 
     // Given: User is logged in
     await loginPage.open();
     await loginPage.login();
-    // And Navigate to the tag page
+    // And Navigate to create tag page
     await tagPage.open();
-    // When I fill the tag name
-    await tagPage.fillTagName('Test Tag');
 
-    async function getSaveTagResponse() {
-        const responsePromise = await page.waitForResponse(async (response) => {
-            if (!response.url().includes('tags')) return false;
-            return response.status() === 201 || response.status() === 200;
-        });
-        return responsePromise.json();
-    }
+    // When: I fill the tag description with more than 500 characters
+    await tagPage.fillTagDescription('a'.repeat(501));
     // And I save the tag
     await tagPage.saveTag();
-    const response = await getSaveTagResponse();
     await takeScreenshot(page);
-    // Then It should create the tag and return the name
-    expect(response.tags[0].name).toBe('Test Tag');
+    // Then It should show an error
+    const error = await tagPage.getSaveFailure();
+    expect(await error.isVisible()).toBeTruthy();
+    expect(await error.innerText()).toBe('Retry');
 });
