+++ features/create_tag_with_invalid_description.feature.commented
@@ -1,4 +1,4 @@
-Feature: Create tag with invalid description EP020
+Feature: Create tag with invalid description EP018
 
   @user1 @web
   Scenario: As user I login and create a tag with invalid description.
